fix(tests): use portable ioctl request type

+++ tests/tui_smoke.rs
@@ -452,7 +452,7 @@ fn spawn_dext_with_env(
             if libc::setsid() == -1 {
                 return Err(io::Error::last_os_error());
             }
-            if libc::ioctl(libc::STDIN_FILENO, libc::TIOCSCTTY, 0) == -1 {
+            if libc::ioctl(libc::STDIN_FILENO, libc::TIOCSCTTY as libc::c_ulong, 0) == -1 {
                 return Err(io::Error::last_os_error());
             }
             Ok(())
